+++ example/servo.js
@@ -1,0 +1,18 @@
+'use strict';
+
+var Gpio = require('../'),
+  motor = new Gpio(10, {mode: Gpio.OUTPUT}),
+  pulseWidth = 1000,
+  increment = 100;
+
+setInterval(function () {
+  motor.servoWrite(pulseWidth);
+
+  pulseWidth += increment;
+  if (pulseWidth >= 2000) {
+    increment = -100;
+  } else if (pulseWidth <= 1000) {
+    increment = 100;
+  }
+}, 1000);
+